Remove HDCP check on external displays
Pixel devices are still enforced, but the check for other devices
is removed to maintain backward compatibility. This needs to be
reverted when HDCP aidl API becomes ready.

This is safe (and unflagged) because it simply restores the
long-tested Android U behavior for non-Pixel devices, while
leaving the Pixel behavior (where mIsHdcpViaNegVsync is already
set to true) unchanged. This also needs to be cherry-picked to
Android 15, where it cannot be flagged.

Flag: EXEMPT bugfix
Bug: 347825589
Test: play DRM protected video with external display connected
Change-Id: Iec423d9b0d127474c734beb8a0d73f9ce780a4e2