Add more dispatcher tests for appearing windows
It has been suggested that a window that appears while the gesture is
active may cause incorrect behaviour in dispatcher if it then touched
with additional new pointers.

These tests attempt to check some of those scenarios. Currently, there's
no evidence that this could lead to a crash.

Flag: TEST_ONLY
Bug: 341869464
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_break_on_failure --gtest_filter="*SplitTouches*"
Change-Id: I856907958684e84d08858aa4b92199a28095e58a