libbinder: export fewer symbols for non-vendor builds
We did this for bionic builds only in https://r.android.com/3096103, but
it was reverted in https://r.android.com/3148521 because of issues in
vendor code. We no longer need to limit this to bionic builds, so I've
simplified that part of the change and added an override to disable for
vendor builds.

Bug: 338458975
Bug: 349657329
Test: boot mokey without seeing a libbinder related crash
Change-Id: I6a67b4aea73ce496b2460a084736c472ed5e2430