Remove some logging information from InputDispatcher
Currently, we are often printing information like "edgeFlags" and
"xPrecision"/"yPrecision". Over the years, I've never found that
information useful, and it clutters the logs.

Separately, we are labeling input channels like "(server) and
"(client)", which is also never useful. After the input channels were
refactored several years ago, it's pretty hard to mix those up:
dispatcher always has a unique_ptr of the InputChannel, so there's
little risk of it leaking elsewhere.

In this CL, those pieces of logs are removed to make it easier to read
the various dumps and logs.

Bug: 347700797
Flag: EXEMPT refactor
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_break_on_failure --gtest_filter="*SpyThatPreventsSplittingWithApplication*"
Change-Id: I7414a6c35a7a1bc1c779d51cd7243b30356f803e