Dump dispatcher state during publisher error
When publisher encounters an error, it crashes. Today, those crashes
lack the corresponding dispatcher context.

In this CL, we change the pipeline so that the dispatcher crashes
instead. This would help us see the dump of the dispatcher state and get
better idea of how things went wrong.

Bug: 347700797
Flag: EXEMPT refactor
Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_break_on_failure --gtest_filter="*SpyThatPreventsSplittingWithApplication*"
Change-Id: Iac148751703de833503c2638ef20dcb4cb137297