Skip logging error when service name doesn't match vintf format
VINTF services are expected to have a specific format for their names.
We now build an AidlName for each service so we can check for them in
the manifest files. In these cases we don't want to produce a log when
the format doesn't match because these services are not all vintf
services.

Test: m
Bug: 355053763
Change-Id: I76a12985495892809d3319963a9c020d60fd4084