Set down time when event slips into Wallpaper
When touch slips into a window that also has a wallpaper, the wallpaper
should receive a slippery enter event, as well. We should also ensure
that the downTime is correctly set at this time. Otherwise, we will
encounter issues later, when trying to split motion events.

This can be reproduced by slipping touch from the navigationbar into
launcher and then pressing the navigation bar with other finger.

Bug: 354059964
Test: manually
1. Press navigationbar and move to launcher and hold.
2. Inject motionevent by "adb shell input swipe 500 1500 500 1000"
3. Press navigationbar by other finger.

Test: TEST=inputflinger_tests; m $TEST && $ANDROID_HOST_OUT/nativetest64/$TEST/$TEST --gtest_filter="*WallpaperWindowWhenSlipperyAndMultiWindowMultiTouch"
Flag: EXEMPT bugfix

Change-Id: Ib201e859492509a6cd9e7f73065813ef15611ad9