evemu-record: rename "boot" timestamp base to "epoch"
My intention when I implemented this feature was that timestamps would
be made relative to when the system booted, but I was mistaken about
what timestamp base the Kernel events used and didn't test it thoroughly
enough, as the feature ended up outputting Unix timestamps (i.e.
relative to the epoch). This actually turns out to be more useful when
matching the evemu recordings to other logs (since you can just turn the
timestamps into dates and times directly, rather than having to work out
when the system booted). So, let's just rename the command-line option
and pretend this was what I intended all along...

Bug: 343942561
Test: run `evemu-record --timestamp-base=epoch`, and check the
      timestamps in the output by comparing them to `date +%s`
Flag: EXEMPT dev tool not used in any automation
Change-Id: I80e3b4db1bcc2e6f656bfaaf7904f1678cfb29cf