libbinder: allow libbinder_ndk + libbinder.vendor
Due to messed up fuzzing linker namespaces, which are being fixed,
when fuzzers are configured as:
   - vendor: true,
   - shared_libs: [... "libbinder_ndk" ...],

then the runtime libraries used are like this:
   - /data/fuzz/x86_64/lib/vendor/libbinder.so
   - /system/lib64/libbinder_ndk.so

This is not a correct configuration. However, we can change the
way openDeclaredPassthroughHal is disabled in order to provide
the same ELF-level interface (that libbinder_ndk expects) from
libbinder as libbinder.vendor.

Future considertions:
- very soon, we need to remove this implicit requirement on the
  dependencies of ll-ndk libraries, by fixing fuzzer linker
  namespaces
- vendors should be able to dlopen same-process stable C HALs, as
  system can, but this requires additional apexd APIs available
  to vendor

Bugs: me
Test: run all service fuzzers, no startup crashes
Change-Id: Icf09de3ca048b542d7c22e6f4ca030977be3dff4