Adding an InputEventActionType enum field to InputEventTimeline.
This is needed for the Per Device Input Latency metrics which will be implemented, because we would
like to filter the data per input event action type. This enum corresponds to the InputEventType
enum field that will be added to the InputEventLatency atom.

InputEventTimeline already had a bool field isDown that was true only if the Motion Event was an
ACTION_DOWN, so we were dividing the latency values into two categories: DOWN vs other types of
events. The isDown field is still needed for the InputEventLatencySketch atom, but should be
deprecated with this atom once the new atom is recording data.

Bug: b/270049345
Test: atest inputflinger_tests
Flag: EXEMPT bugfix

Change-Id: Iabccb7217820d4ed070378d8f4d8f0410a2efbdf