Fix C compatibility in new API.
I fixed all these in AOSP, but this API isn't in AOSP yet and is also
missing the typedef needed for C compat.

Also removed the duplicate decl, and the unnecessary __INTRODUCED_IN
that either does nothing for types, or is actively harmful (pretty
sure it does nothing), and the accompanying incorrect "introduced in"
docs.

Flag: no
Bug: None
Test: treehugger
Change-Id: I9432f1146288dcb691e5f317887791cc94b120fe