Use different timeouts based on whether event is expected to occur
Before this change, the test waited for 500 ms in the cases where the
input devices were expected to change, and also in the cases where the
input devices were not expected to change.

When a device is being added, this timeout is way too short for some
hardware. This causes test flakiness. However, we can't directly
increase this timeout, because it would make the tests slower, since the
same value will be used for the case where input devices are not
expected to change.

Therefore, in this CL, we provide a way for the caller to specify the
timeout, and introduce a new, shorter timeout for the cases where the
event is not expected to occur.

This is a speculative fix.

Bug: 332178280
Change-Id: I6ee5b75a1331c41544b3efa4c408964eeb512551
Flag: TEST_ONLY
Test: atest inputflinger_tests