Don't affect window focus on touchpad gesture
When using a multi-touch touchpad, it is the expected behavior in most
operating systems that the user is allowed to perform gestures (like
scroll, pinch, etc.) on an unfocused window without bringing it into
focus.

This change adds the NO_FOCUS_CHANGE flag to the MotionEvents generated
by GestureConverter. The behavior was initially added in commit [1] in
TouchInputMapper, but got lost when touchpad stack was introduced.

[1] I74e52f8daa13d4e6c047bc23982ec56942c555f6

Bug: 364460018
Test: GestureConverterTest
Test: Manually open two windows on desktop mode, and see if gesture won't change focus
Flag: com.android.input.flags.enable_touchpad_no_focus_change
Change-Id: I370189260c96965698ee2d0885edf32378dbca74