TestEventMatchers: bounds check pointer indexes
b/365166534 was caused by a WithPointer… matcher not checking that its
pointer index was in bounds, and this check is missing from a couple of
other WithPointer… matchers. Add it, so that more bugs like this can be
found instantly in future rather than causing flaky assertions against
uninitialized memory.

Test: atest inputflinger_tests
Bug: 365166534
Flag: TEST_ONLY
Change-Id: I57d23909888749e15d4d034ddb0c091336b016d1