PointerChoreographer: Show spot hover icon for stylus when showing taps
When "show touches" is enabled, we show the spot hover icon when the
stylus is hovering. However, when an app is allowed to override the icon
for a hovering stylus, we were never showing the spot hover icon.

In this CL, we ensure that the spot hover icon is visible for a hovering
stylus when "show touches" is enabled and the app has not specified an
icon to use.

Bug: 342067371
Test: atest inputflinger_tests
Flag: EXEMPT bug fix
Change-Id: Iddf637c87ecba928a5a227a8efc0dd436b9528c7